+++ data/queries.txt
@@ -1,0 +1,60 @@
+1 world interest rates table
+2 2008 beijing olympics
+3 fast cars
+4 clothing sizes
+5 phases of the moon
+6 usa population by state
+7 prime ministers of england
+8 ipod models
+9 bittorrent clients
+10 olympus digital slrs
+11 composition of the sun
+12 running shoes
+13 fuel consumption
+14 stock quote tables
+15 top grossing movies
+16 nutrition values
+17 state capitals and largest cities in us
+18 professional wrestlers
+19 company income statements
+20 dog breeds
+21 ibanez guitars
+22 used cellphones
+23 world religions
+24 stocks
+25 academy awards
+26 2008 olympic gold medal winners
+27 currencies of different countries
+28 science discoveries
+29 pga leaderboard
+30 pain medications
+31 football clubs city
+32 healthy food cost
+33 capitals attractions
+34 diseases mortality
+35 cigarette brands market share
+36 apples market share
+37 healthy food nutritional value
+38 hormones effects
+39 household chemicals strength
+40 lakes altitude
+41 laptops cpu
+42 asian countries currency
+43 diseases risks
+44 external drives capacity
+45 baseball teams captain
+46 maryland counties population
+47 countries capital
+48 diseases incidence
+49 eu countries year joined
+50 irish counties area
+51 cereals nutritional value
+52 erp systems price
+53 cats life span
+54 broadway musicals director
+55 infections treatment
+56 food type
+57 board games number of players
+58 google products reviews
+59 constellations closest constellation
+60 games age
